Update for GCC 14.1.0 (aarch64).
The major change here is the building of gdb.

In the past, because the native compiler wasn't intended to be
relocated, we could pick up libgmp.dylib, libmpfr.dylib, and
libstdc++.dylib from the native compiler (actually, the native
compiler doesn't now include libmpfr, so that had to change already).

We can resolve the libgmp.dylib problem in the same way as we'd
already resloved the libmpfr.dylib problem, namely by making static
builds.

To resolve the libstdc++.dylib problem, we build with clang and
clang++, so we pick up Apple's standard libc++.dylib.

  * Makefile (gmp): new.
    (mpfr): depends on gmp.
  * common.sh: shares some design with the equivalent for native builds.
    (VERSION): defaults to 14.1.0.
    (PYTHON): new, python3.9.
    (CORES), new.
    (PATH): set to the place where the native compiler was built.
    (PREFIX): uses $ARCH.
    (GCC_SRC): several choices, defaults to FSF GCC.
    (NEW_PATH): add the new cross-compiler to the front of the path.
      Not actually used, could be used in newlib.sh instead of extant
      setting.
  * binutils.sh: make with -j$CORES.
  * gcc-boot.sh: likewise.
  * gcc.sh: likewise.
  * newlib.sh: likewise.
  * gdb.sh: pick up GMP from $PREFIX. Build with clang, clang++. Make
      with -j$CORES.
  * gmp.sh: new.
  * mpfr.sh: pick up GMP from $PREFIX. Make with -j$CORES.
  * README.md: simplified by removing references to GNAT GPL builds and
      by referring to the Makefile.
  * release_notes.md: as summary.